Prueba tiempos de validación con metars usando tres scripts distintos

+++ METAR/ecmwf_valida_metar_parseacsv.sh
@@ -108,32 +108,28 @@ fi
 echo "$(datePID): Inicia descarga de METAR"
 curl -s -m $LIM --connect-timeout $TIMEOUT http://aire.ogimet.com/meteoflight_reports.php -o $filemetarID
 ##Descarga METAR info
-curl -s -m $LIM --connect-timeout $TIMEOUT "http://www.ogimet.com/cgi-bin/getmetar?begin=${fechaini}00&end=${fechafin}00" -o $fnameMETAR
+#curl -s -m $LIM --connect-timeout $TIMEOUT "http://www.ogimet.com/cgi-bin/getmetar?begin=${fechaini}00&end=${fechafin}00" -o $fnameMETAR
+curl -s -m $LIM --connect-timeout $TIMEOUT "http://www.ogimet.com/cgi-bin/getmetar?icao=LE&begin=${fechaini}00&end=${fechafin}00" -o $fnameMETAR
 echo "$(datePID): Fin descarga de METAR"
 
-metarID=$(cat "$filemetarID" | jq '.points[].icao' | grep '[L][E][A-Z][A-Z]' | cut -c2-5)
-
-echo $metarID > ppMETARs
-#Truco de Juan para asegurarme que tengo t-odos los json descargados adecaudamente
-
-#loop sobre los METARs para descargar la predicción y comparar
-for icao in $metarID; do
-    fnamePREDIC=${DIR_DATA}/PREDIC-$icao-$fechaini-$fechafin
-    fnameVALIDA=${DIR_DATA}/VALIDA-$icao-$fechaini-$fechafin
-    fnameECMWF=${DIR_DATA}/ECMWF-$icao-${weekDOWNLOAD} #creado con get-info-ECMWF.sh
-
-      echo "$(datePID): Ini descarga de json"
-      ##Descarga forecast info del json file en la localidad del METAR
-      lat=$(cat $filemetarID | jq '[.points[] | select(.icao == "'$icao'") | .lat] | sort []')
-      lon=$(cat $filemetarID | jq '[.points[] | select(.icao == "'$icao'") | .lon] | sort []')
-      alt=$(cat $filemetarID | jq '[.points[] | select(.icao == "'$icao'") | .alt] | sort []')
-      curl -s -m $LIM --connect-timeout $TIMEOUT 'http://aguila.ogimet.com/cgi-bin/otf12?latitud='$lat'&longitud='$lon'&altitud='$alt'&zonaHoraria=etc/UTC&name='$icao'' -o $fnamePREDIC.json
-      echo "$(datePID): Fin descarga de json"
-      cat $fnameMETAR | grep "${icao}" | grep '[ M][0-9][0-9]/[M]*[0-9][0-9]' > kk$icao
       tempSUMBIAS=0
       tempSUMRMSE=0
-      while IFS=, read -r metar YYYY MM DD HH mm infoMETAR
+while IFS=, read -r icao YYYY MM DD HH mm infoMETAR
       do
+        fnamePREDIC=${DIR_DATA}/PREDIC-$icao-$fechaini-$fechafin
+        fnameVALIDA=${DIR_DATA}/VALIDA-$icao-$fechaini-$fechafin
+        fnameECMWF=${DIR_DATA}/ECMWF-$icao-${weekDOWNLOAD} #creado con get-info-ECMWF.sh
+
+        if [ ! -e "$fnamePREDIC".json ]; then
+          echo "$(datePID): Ini descarga de json"
+          echo "no existe $fnamePREDIC" >> kkpruebacondicional
+          ##Descarga forecast info del json file en la localidad del METAR
+          lat=$(cat $filemetarID | jq '[.points[] | select(.icao == "'$icao'") | .lat] | sort []')
+          lon=$(cat $filemetarID | jq '[.points[] | select(.icao == "'$icao'") | .lon] | sort []')
+          alt=$(cat $filemetarID | jq '[.points[] | select(.icao == "'$icao'") | .alt] | sort []')
+          curl -s -m $LIM --connect-timeout $TIMEOUT 'http://aguila.ogimet.com/cgi-bin/otf12?latitud='$lat'&longitud='$lon'&altitud='$alt'&zonaHoraria=etc/UTC&name='$icao'' -o $fnamePREDIC.json
+          echo "$(datePID): Fin descarga de json"
+        fi
         fechaMETAR=$(date -d "${YYYY}/${MM}/${DD}H${HH}:${mm}" "+%s")
         fechaMETARok=$(echo "${YYYY}${MM}${DD}${HH}${mm}")
 
@@ -144,15 +140,22 @@ for icao in $metarID; do
         if [ ! -z "$fechaPRED" ] #puede ocurrir que la fecha del METAR no esté en el json de la predicc
         then
           tempPRED=$(jq '.dias[].horas[] | select (.utime == '$fechaPRED') | .temperatura.valor' $fnamePREDIC.json)
-          #tempMETAR=$(echo $infoMETAR | grep "${icao}" | grep -o '[ M][0-9][0-9]/[M]*[0-9][0-9]' | tr "M" "-" | cut -c1-3)
-          tempMETAR=$(echo $infoMETAR | grep "${metar}" | grep -o '[ M][0-9][0-9]/[M]*[0-9][0-9]'| tr "M" "-" | tr "/" " " | awk '{print $1}')
-          #mmddhhii=
+          tempMETAR=$(echo $infoMETAR | grep "${icao}" | grep -o '[ M][0-9][0-9]/[M]*[0-9][0-9]'| tr "M" "-" | tr "/" " " | awk '{print $1}')
+           #TODO incluir datos del ECMWF
+          #mmddhhii=fechaPRED
           #grib=/home/ecmwf/${cc}${S}${MMDDHHII}${mmddhhii}1
           #if [ -e $grib ]
           #then
           #  else
           #    tempECMWF='NAN'
           #fi
+        tempECMWF="NAN"
+        tempBIAS=$(echo $tempPRED - $tempMETAR | bc)
+        tempRMSE=$(echo ${tempBIAS}*${tempBIAS}  | bc)
+
+        tempSUMBIAS=$(echo $tempSUMBIAS + $tempBIAS | bc)
+        tempSUMRMSE=$(echo $tempSUMRMSE + $tempRMSE | bc)
+
           #Nos aseguramos que los datos se comparan con la última pasada
           if [ -e $fnameVALIDA.txt ]
           then
@@ -161,22 +164,21 @@ for icao in $metarID; do
             rm kkrmdate
           fi
         fi
-        #TODO incluir datos del ECMWF
-        tempECMWF="NAN"
-        tempBIAS=$(echo $tempPRED - $tempMETAR | bc)
-        tempRMSE=$(echo ${tempBIAS}*${tempBIAS}  | bc)
-
-        tempSUMBIAS=$(echo $tempSUMBIAS + $tempBIAS | bc)
-        tempSUMRMSE=$(echo $tempSUMRMSE + $tempRMSE | bc)
 
         # Escribe datos en un archivo .txt para que pinte R
         echo "$fechaMETARok" "$fechaPRED" "$tempPRED" "$fechaMETAR" "${tempMETAR}" "$tempECMWF" "$tempSUMBIAS" "$tempSUMRMSE" >> ${fnameVALIDA}.txt
-      done<kk$icao
-      echo "$(datePID):Fin parseo fechas del $icao"
+done<"${fnameMETAR}"
+
+echo "$(datePID):Fin parseo fechas del csv"
+exit
+#TODO pensar otra manera para quitarse el loop sobre los icao
+metarID=$(cat "$filemetarID" | jq '.points[].icao' | grep '[LE][A-Z][A-Z][A-Z]' | cut -c2-5)
+for icao in $metarID; do
       valida=$(cat -n ${fnameVALIDA}.txt | tail -1)
-      echo "${icao}" "$lon" "$lat" "$alt" $valida | awk '{print $1,$2,$3,$12/$5,sqrt($13)/$5}' | awk 'NF==5' >> ${DIR_BASE}/VALIDA-mundo-$fechaini-$fechafin.txt
-      rm $fnamePREDIC.json kk$icao
-done #loop del icao
+      echo "icao" "$lon" "$lat" "$alt" $valida | awk '{print $1,$2,$3,$12/$5,sqrt($13)/$5}' | awk 'NF==5' >> ${DIR_BASE}/VALIDA-mundo-$fechaini-$fechafin.txt
+      rm $fnamePREDIC.json
+done
+
 echo "$(datePID):Fin parseo loop ICAO"
   #warning: si se cambian los parámetros de entradas modificar R consecuentemente
   filein=${DIR_BASE}/VALIDA-mundo-$fechaini-$fechafin.txt
@@ -206,7 +208,6 @@ R --slave --args "${filein}" "${plotout}" << EOF
             panel.background = element_blank(), panel.border = element_rect(linetype = "solid", fill = NA))
 EOF
 echo "$(datePID):Fin pintar mapa"
-#  Rscript --vanilla $scriptR1 "$filein" "$plotout"
 
  # rm $fnameMETAR $filemetarID
 echo "Fin descarga ecmwf data"
